requirements.txt
transformers
streamlit
torch
huggingface-hub
python-dotenv